requirements.txt
langchain-openai>=0.1.0
langchain-core>=0.1.0
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
nest-asyncio>=1.5.0

